requirements.txt
adbc-driver-manager==0.11.0
adbc-driver-postgresql==0.11.0
annotated-types==0.6.0
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
duckdb==0.10.2
filelock==3.13.4
gdown==5.1.0
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
httplib2==0.22.0
idna==3.7
importlib_resources==6.4.0
multimethod==1.10
mypy-extensions==1.0.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pandera==0.18.3
polars==0.20.22
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyarrow==16.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
pytz==2024.1
requests==2.31.0
rsa==4.9
six==1.16.0
soupsieve==2.5
SQLAlchemy==1.4.36
tqdm==4.66.2
typeguard==4.2.1
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
wrapt==1.16.0
